LRDOCS-9604 small fixes
