--- docs/dxp/latest/en/site-building/site-navigation/developer-guide/navigation-menu-api-basics/resources/curl/NavigationMenu_PUT_ById.sh
@@ -1,6 +1,0 @@
-curl \
-	-H "Content-Type: application/json" \
-	-X PUT \
-	"http://localhost:8080/o/headless-delivery/v1.0/navigation-menus/${1}" \
-	-d "{\"name\": \"Goo\"}" \
-	-u "test@liferay.com:test"
