--- docs/dxp/latest/en/site-building/site-navigation/developer-guide/navigation-menu-api-basics/resources/java/NavigationMenus_GET_FromSite.java
@@ -1,27 +1,0 @@
-import com.liferay.headless.delivery.client.dto.v1_0.NavigationMenu;
-import com.liferay.headless.delivery.client.pagination.Page;
-import com.liferay.headless.delivery.client.pagination.Pagination;
-import com.liferay.headless.delivery.client.resource.v1_0.NavigationMenuResource;
-
-public class NavigationMenus_GET_FromSite {
-
-	/**
-	 * java -classpath .:* -DsiteId=1234 NavigationMenus_GET_FromSite
-	 */
-	public static void main(String[] args) throws Exception {
-		NavigationMenuResource.Builder builder =
-			NavigationMenuResource.builder();
-
-		NavigationMenuResource navigationMenuResource = builder.authentication(
-			"test@liferay.com", "test"
-		).build();
-
-		Page<NavigationMenu> page =
-			navigationMenuResource.getSiteNavigationMenusPage(
-				Long.valueOf(System.getProperty("siteId")), null, null, null,
-				null, Pagination.of(1, 2), null);
-
-		System.out.println(page);
-	}
-
-}
