+++ docs/dxp/latest/en/site-building/site-navigation/developer-guide/navigation-menu-api-basics/resources/liferay-p7s4.zip/curl/NavigationMenu_POST_ToSite.sh
