+++ docs/dxp/latest/en/site-building/site-navigation/developer-guide/navigation-menu-api-basics/resources/liferay-p7s4.zip/curl/NavigationMenu_PUT_ById.sh
@@ -1,6 +1,6 @@
 curl \
 	-H "Content-Type: application/json" \
-	-X PATCH \
+	-X PUT \
 	"http://localhost:8080/o/headless-delivery/v1.0/navigation-menus/${1}" \
 	-d "{\"name\": \"Bar\"}" \
 	-u "test@liferay.com:test"
