+++ docs/dxp/latest/en/site-building/site-navigation/developer-guide/navigation-menu-api-basics/resources/liferay-p7s4.zip/curl/NavigationMenus_GET_FromSite.sh
