+++ docs/dxp/latest/en/site-building/site-navigation/developer-guide/navigation-menu-api-basics/resources/update_example.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+source $(git rev-parse --show-toplevel)/_common.sh
+
+download_nexus_jar "com.liferay.headless.delivery.client"
